Simplify `fromZonedTime`
Problem: Currently, the `fromZonedTime` constructors enforce that the
`ZonedTime`'s offset must be a valid offset in the given time zone. If
it's not, we throw/return an error.

However, we don't really _have to_ enforce this rule, or throw/return
any errors. A `ZonedTime`, just like `UTCTime`, represents a valid and
unique point in time.

Solution: Convert the `ZonedTime` to `UTCTime`, and then to `TZTime`.